# Conflicts:
#	Cargo.lock
#	applications/tari_app_grpc/proto/base_node.proto
#	applications/tari_app_grpc/proto/types.proto
#	applications/tari_app_grpc/proto/wallet.proto
#	applications/tari_app_grpc/src/conversions/transaction.rs
#	applications/tari_base_node/src/grpc/base_node_grpc_server.rs
#	applications/tari_console_wallet/src/automation/command_parser.rs
#	applications/tari_console_wallet/src/automation/commands.rs
#	applications/tari_console_wallet/src/grpc/wallet_grpc_server.rs
#	applications/tari_console_wallet/src/ui/app.rs
#	applications/tari_console_wallet/src/ui/components/receive_tab.rs
#	applications/tari_console_wallet/src/ui/state/app_state.rs
#	applications/tari_console_wallet/src/ui/state/wallet_event_monitor.rs
#	base_layer/core/src/blocks/genesis_block.rs
#	base_layer/core/src/transactions/helpers.rs
#	base_layer/core/src/transactions/transaction/mod.rs
#	base_layer/core/src/transactions/transaction_protocol/recipient.rs
#	base_layer/core/src/transactions/transaction_protocol/sender.rs
#	base_layer/core/src/transactions/transaction_protocol/sender_transaction_protocol_builder.rs
#	base_layer/wallet/src/base_node_service/config.rs
#	base_layer/wallet/src/output_manager_service/handle.rs
#	base_layer/wallet/src/output_manager_service/mod.rs
#	base_layer/wallet/src/output_manager_service/service.rs
#	base_layer/wallet/src/output_manager_service/storage/database/mod.rs
#	base_layer/wallet/src/output_manager_service/storage/models.rs
#	base_layer/wallet/src/output_manager_service/storage/sqlite_db/mod.rs
#	base_layer/wallet/src/testnet_utils.rs
#	base_layer/wallet/src/transaction_service/error.rs
#	base_layer/wallet/src/transaction_service/handle.rs
#	base_layer/wallet/src/transaction_service/protocols/transaction_broadcast_protocol.rs
#	base_layer/wallet/src/transaction_service/protocols/transaction_coinbase_monitoring_protocol.rs
#	base_layer/wallet/src/transaction_service/protocols/transaction_receive_protocol.rs
#	base_layer/wallet/src/transaction_service/protocols/transaction_send_protocol.rs
#	base_layer/wallet/src/transaction_service/protocols/transaction_validation_protocol.rs
#	base_layer/wallet/src/transaction_service/service.rs
#	base_layer/wallet/src/transaction_service/storage/database.rs
#	base_layer/wallet/src/transaction_service/storage/sqlite_db.rs
#	base_layer/wallet/src/utxo_scanner_service/utxo_scanning.rs
#	base_layer/wallet/src/wallet.rs
#	base_layer/wallet_ffi/src/lib.rs
#	common/logging/log4rs_sample_base_node.yml
#	common/src/configuration/bootstrap.rs
#	common/src/configuration/global.rs
#	common/src/configuration/network.rs
#	comms/src/connection_manager/wire_mode.rs

+++ integration_tests/features/support/world.js
@@ -3,25 +3,28 @@ const { setWorldConstructor, After, BeforeAll } = require("cucumber");
 const BaseNodeProcess = require("../../helpers/baseNodeProcess");
 const MergeMiningProxyProcess = require("../../helpers/mergeMiningProxyProcess");
 const WalletProcess = require("../../helpers/walletProcess");
+const WalletFFIClient = require("../../helpers/walletFFIClient");
 const MiningNodeProcess = require("../../helpers/miningNodeProcess");
+const TransactionBuilder = require("../../helpers/transactionBuilder");
 const glob = require("glob");
 const fs = require("fs");
 const archiver = require("archiver");
-
 class CustomWorld {
   constructor({ attach, parameters }) {
     // this.variable = 0;
     this.attach = attach;
-
+    this.checkAutoTransactions = true;
     this.seeds = {};
     this.nodes = {};
     this.proxies = {};
     this.miners = {};
     this.wallets = {};
+    this.walletsFFI = {};
     this.walletPubkeys = {};
     this.clients = {};
     this.headers = {};
     this.outputs = {};
+    this.transactionOutputs = {};
     this.testrun = `run${Date.now()}`;
     this.lastResult = null;
     this.blocks = {};
@@ -29,25 +32,26 @@ class CustomWorld {
     this.peers = {};
     this.transactionsMap = new Map();
     this.resultStack = [];
-    this.tipHeight = 0;
     this.logFilePathBaseNode =
       parameters.logFilePathBaseNode || "./log4rs/base_node.yml";
     this.logFilePathProxy = parameters.logFilePathProxy || "./log4rs/proxy.yml";
-    this.logFilePathMiningNocde =
-      parameters.logFilePathMiningNocde || "./log4rs/mining_node.yml";
+    this.logFilePathMiningNode =
+      parameters.logFilePathMiningNode || "./log4rs/mining_node.yml";
     this.logFilePathWallet =
       parameters.logFilePathWallet || "./log4rs/wallet.yml";
   }
 
   async createSeedNode(name) {
+    console.log(`seed:`, name);
     const proc = new BaseNodeProcess(
       `seed-${name}`,
+      false,
       null,
       this.logFilePathBaseNode
     );
     await proc.startNew();
     this.seeds[name] = proc;
-    this.clients[name] = proc.createGrpcClient();
+    this.clients[name] = await proc.createGrpcClient();
   }
 
   seedAddresses() {
@@ -60,7 +64,7 @@ class CustomWorld {
 
   /// Create but don't add the node
   createNode(name, options) {
-    return new BaseNodeProcess(name, options, this.logFilePathBaseNode);
+    return new BaseNodeProcess(name, false, options, this.logFilePathBaseNode);
   }
 
   async createAndAddNode(name, addresses) {
@@ -71,12 +75,12 @@ class CustomWorld {
       node.setPeerSeeds([addresses]);
     }
     await node.startNew();
-    this.addNode(name, node);
+    await this.addNode(name, node);
   }
 
-  addNode(name, process) {
+  async addNode(name, process) {
     this.nodes[name] = process;
-    this.clients[name] = process.createGrpcClient();
+    this.clients[name] = await process.createGrpcClient();
   }
 
   addMiningNode(name, process) {
@@ -87,23 +91,81 @@ class CustomWorld {
     this.proxies[name] = process;
   }
 
-  async createAndAddWallet(name, nodeAddresses) {
-    const wallet = new WalletProcess(name, {}, this.logFilePathWallet);
+  async createAndAddWallet(name, nodeAddresses, options = {}) {
+    const wallet = new WalletProcess(
+      name,
+      false,
+      options,
+      this.logFilePathWallet
+    );
     wallet.setPeerSeeds([nodeAddresses]);
     await wallet.startNew();
 
     this.addWallet(name, wallet);
-    let walletClient = wallet.getClient();
+    let walletClient = await wallet.connectClient();
     let walletInfo = await walletClient.identify();
     this.walletPubkeys[name] = walletInfo.public_key;
+  }
+
+  async createAndAddFFIWallet(name, seed_words = null, passphrase = null) {
+    const wallet = new WalletFFIClient(name);
+    await wallet.startNew(seed_words, passphrase);
+    this.walletsFFI[name] = wallet;
+    this.walletPubkeys[name] = wallet.identify();
+    return wallet;
   }
 
   addWallet(name, process) {
     this.wallets[name] = process;
   }
 
+  addWalletPubkey(name, pubkey) {
+    this.walletPubkeys[name] = pubkey;
+  }
+
   addOutput(name, output) {
     this.outputs[name] = output;
+  }
+
+  addTransactionOutput(spendHeight, output) {
+    if (this.transactionOutputs[spendHeight] == null) {
+      this.transactionOutputs[spendHeight] = [output];
+    } else {
+      this.transactionOutputs[spendHeight].push(output);
+    }
+  }
+
+  async createTransactions(name, height) {
+    let result = true;
+    const txInputs = this.transactionOutputs[height];
+    if (txInputs == null) {
+      return result;
+    }
+    let i = 0;
+    for (const input of txInputs) {
+      const txn = new TransactionBuilder();
+      txn.addInput(input);
+      const txOutput = txn.addOutput(txn.getSpendableAmount());
+      this.addTransactionOutput(height + 1, txOutput);
+      const completedTx = txn.build();
+      const submitResult = await this.getClient(name).submitTransaction(
+        completedTx
+      );
+      if (this.checkAutoTransactions && submitResult.result != "ACCEPTED") {
+        result = false;
+      }
+      if (submitResult.result == "ACCEPTED") {
+        i++;
+      }
+      if (i > 9) {
+        //this is to make sure the blocks stay relatively empty so that the tests don't take too long
+        break;
+      }
+    }
+    console.log(
+      `Created ${i} transactions for node: ${name} at height: ${height}`
+    );
+    return result;
   }
 
   async mineBlock(name, weight, beforeSubmit, onError) {
@@ -114,10 +176,15 @@ class CustomWorld {
     );
   }
 
-  baseNodeMineBlocksUntilHeightIncreasedBy(baseNode, wallet, numBlocks) {
+  async baseNodeMineBlocksUntilHeightIncreasedBy(baseNode, wallet, numBlocks) {
+    let w = null;
+    if (wallet) {
+      let tmp = this.getWallet(wallet);
+      w = await tmp.connectClient();
+    }
     const promise = this.getClient(baseNode).mineBlocksUntilHeightIncreasedBy(
       numBlocks,
-      wallet ? this.getWallet(wallet).getClient() : null
+      w
     );
     return promise;
   }
@@ -156,32 +223,69 @@ class CustomWorld {
   }
 
   getClient(name) {
-    return this.clients[name];
+    const client = this.clients[name];
+    if (!client) {
+      throw new Error(`Node client not found with name '${name}'`);
+    }
+    return client;
   }
 
   getNode(name) {
-    return this.nodes[name] || this.seeds[name];
+    const node = this.nodes[name] || this.seeds[name];
+    if (!node) {
+      throw new Error(`Node not found with name '${name}'`);
+    }
+    return node;
   }
 
   getMiningNode(name) {
-    return this.miners[name];
+    const miner = this.miners[name];
+    if (!miner) {
+      throw new Error(`Miner not found with name '${name}'`);
+    }
+    return miner;
   }
 
   getWallet(name) {
-    return this.wallets[name];
+    const wallet = this.wallets[name] || this.walletsFFI[name];
+    if (!wallet) {
+      throw new Error(`Wallet not found with name '${name}'`);
+    }
+    return wallet;
   }
 
   getWalletPubkey(name) {
     return this.walletPubkeys[name];
   }
 
+  async getNodeOrWalletClient(name) {
+    let client = this.clients[name.trim()];
+    if (client) {
+      client.isNode = true;
+      client.isWallet = false;
+      return client;
+    }
+    let wallet = this.wallets[name.trim()];
+    if (wallet) {
+      let client = await wallet.connectClient();
+      client.isNode = false;
+      client.isWallet = true;
+      return client;
+    }
+    let ffi_wallet = this.walletsFFI[name.trim()];
+    if (ffi_wallet) {
+      return ffi_wallet;
+    }
+    return null;
+  }
+
   async getOrCreateWallet(name) {
-    const wallet = this.getWallet(name);
+    const wallet = this.wallets[name];
     if (wallet) {
       return wallet;
     }
     await this.createAndAddWallet(name, this.seedAddresses());
-    return this.getWallet(name);
+    return this.wallets[name];
   }
 
   getProxy(name) {
@@ -208,17 +312,18 @@ class CustomWorld {
 
     return new Promise((resolve, reject) => {
       for (let promise of promises) {
-        Promise.resolve(promise).then(
-          () => {
+        Promise.resolve(promise)
+          .then(() => {
             succeeded += 1;
             console.log(`${succeeded} of ${total} (need ${total - canFail})`);
             if (succeeded >= total - canFail) resolve();
-          },
-          () => {
+          })
+          .catch((err) => {
+            console.error(err);
             failed += 1;
-            if (failed > canFail) reject("Too many failed.");
-          }
-        );
+            if (failed > canFail)
+              reject(`Too many failed. Expected at most ${canFail} failures`);
+          });
       }
     });
   }
@@ -228,9 +333,9 @@ class CustomWorld {
     await node.stop();
   }
 
-  async startNode(name) {
+  async startNode(name, args) {
     const node = this.seeds[name] || this.nodes[name];
-    await node.start();
+    await node.start(args);
   }
 
   addTransaction(pubKey, txId) {
@@ -268,49 +373,36 @@ BeforeAll({ timeout: 1200000 }, async function () {
     "compile",
     "127.0.0.1:9999",
     null,
-    "127.0.0.1:9998",
-    this.logFilePathMiningNocde
+    "127.0.0.1:9998"
+    // this.logFilePathMiningNode
   );
   console.log("Compiling mining node...");
   await miningNode.init(1, 1, 1, 1, true, 1);
   await miningNode.compile();
 
+  console.log("Compiling wallet FFI...");
+  await WalletFFIClient.Init();
   console.log("Finished compilation.");
 });
 
 After(async function (testCase) {
   console.log("Stopping nodes");
-  for (const key in this.seeds) {
-    if (testCase.result.status === "failed") {
-      await attachLogs(`${this.seeds[key].baseDir}`, this);
-    }
-    await this.stopNode(key);
-  }
-  for (const key in this.nodes) {
-    if (testCase.result.status === "failed") {
-      await attachLogs(`${this.nodes[key].baseDir}`, this);
-    }
-    await this.stopNode(key);
-  }
-  for (const key in this.proxies) {
-    if (testCase.result.status === "failed") {
-      await attachLogs(`${this.proxies[key].baseDir}`, this);
-    }
-    await this.proxies[key].stop();
-  }
-  for (const key in this.wallets) {
-    if (testCase.result.status === "failed") {
-      await attachLogs(`${this.wallets[key].baseDir}`, this);
-    }
-    await this.wallets[key].stop();
-  }
-  for (const key in this.miners) {
-    if (testCase.result.status === "failed") {
-      await attachLogs(`${this.miners[key].baseDir}`, this);
-    }
-    await this.miners[key].stop();
-  }
+  await stopAndHandleLogs(this.seeds, testCase, this);
+  await stopAndHandleLogs(this.nodes, testCase, this);
+  await stopAndHandleLogs(this.proxies, testCase, this);
+  await stopAndHandleLogs(this.wallets, testCase, this);
+  await stopAndHandleLogs(this.walletsFFI, testCase, this);
+  await stopAndHandleLogs(this.miners, testCase, this);
 });
+
+async function stopAndHandleLogs(objects, testCase, context) {
+  for (const key in objects) {
+    if (testCase.result.status === "failed") {
+      await attachLogs(`${objects[key].baseDir}`, context);
+    }
+    await objects[key].stop();
+  }
+}
 
 function attachLogs(path, context) {
   return new Promise((outerRes) => {
@@ -332,7 +424,7 @@ function attachLogs(path, context) {
           fs.createReadStream("./temp/logzip.zip"),
           "application/zip",
           function () {
-            fs.rmSync("./temp/logzip.zip");
+            fs.rmSync && fs.rmSync("./temp/logzip.zip");
             outerRes();
           }
         );
